dashboard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.32.3
